requirements.txt
earthkit-data
earthkit-maps
earthkit-regrid  
cf-units         
polytope-client
ecmwflibs
cfgrib
lxml
conflator==0.1.5
ipykernel
xarray[complete]
requests
aiohttp
colorama
PyJWT==2.1.0
pydantic_settings
PyYAML
flake8
isort
loguru
tqdm
seaborn
